+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 + [x] Validator: fix for issue #1254
 + [x] Inputs: `required`, `invalid`, `valid` classes now works only for inputs.
 + [x] Validator: use attribute `data-use-required-class` to disable or enable class `required` for inputs with `data-validate=required`
++ [x] Popover: fix firing event `onPopoverShow`, issue #1258
++ [x] Notify: fix default options from new notify
 
 ### 4.2.28
 + [x] Sidebar: fix z-index